+++ java-client/src/main/java/org/opensearch/client/opensearch/_types/mapping/FieldType.java
@@ -109,8 +109,6 @@ public enum FieldType implements JsonEnum {
 
     FlatObject("flat_object"),
 
-    Shape("shape"),
-
     XyShape("xy_shape"),
 
     Histogram("histogram"),