+++ java-client/src/main/java/org/opensearch/client/opensearch/_types/query_dsl/Query.java
@@ -142,8 +142,6 @@ public class Query implements TaggedUnion<Query.Kind, Object>, AggregationVarian
 
         ScriptScore("script_score"),
 
-        Shape("shape"),
-
         XyShape("xy_shape"),
 
         SimpleQueryString("simple_query_string"),
@@ -917,23 +915,6 @@ public class Query implements TaggedUnion<Query.Kind, Object>, AggregationVarian
     }
 
     /**
-     * Is this variant instance of kind {@code shape}?
-     */
-    public boolean isShape() {
-        return _kind == Kind.Shape;
-    }
-
-    /**
-     * Get the {@code shape} variant value.
-     *
-     * @throws IllegalStateException
-     *             if the current variant is not of the {@code shape} kind.
-     */
-    public ShapeQuery shape() {
-        return TaggedUnionUtils.get(this, Kind.Shape);
-    }
-
-    /**
      * Is this variant instance of kind {@code xy_shape}?
      */
     public boolean isXyShape() {
@@ -1658,16 +1639,6 @@ public class Query implements TaggedUnion<Query.Kind, Object>, AggregationVarian
             return this.scriptScore(fn.apply(new ScriptScoreQuery.Builder()).build());
         }
 
-        public ObjectBuilder<Query> shape(ShapeQuery v) {
-            this._kind = Kind.Shape;
-            this._value = v;
-            return this;
-        }
-
-        public ObjectBuilder<Query> shape(Function<ShapeQuery.Builder, ObjectBuilder<ShapeQuery>> fn) {
-            return this.shape(fn.apply(new ShapeQuery.Builder()).build());
-        }
-
         public ObjectBuilder<Query> simpleQueryString(SimpleQueryStringQuery v) {
             this._kind = Kind.SimpleQueryString;
             this._value = v;
@@ -1877,7 +1848,6 @@ public class Query implements TaggedUnion<Query.Kind, Object>, AggregationVarian
         op.add(Builder::regexp, RegexpQuery._DESERIALIZER, "regexp");
         op.add(Builder::script, ScriptQuery._DESERIALIZER, "script");
         op.add(Builder::scriptScore, ScriptScoreQuery._DESERIALIZER, "script_score");
-        op.add(Builder::shape, ShapeQuery._DESERIALIZER, "shape");
         op.add(Builder::simpleQueryString, SimpleQueryStringQuery._DESERIALIZER, "simple_query_string");
         op.add(Builder::spanContaining, SpanContainingQuery._DESERIALIZER, "span_containing");
         op.add(Builder::fieldMaskingSpan, SpanFieldMaskingQuery._DESERIALIZER, "field_masking_span");