+++ java-client/src/main/java/org/opensearch/client/opensearch/_types/query_dsl/QueryBuilders.java
@@ -344,14 +344,7 @@ public class QueryBuilders {
     }
 
     /**
-     * Creates a builder for the {@link ShapeQuery shape} {@code Query} variant.
-     */
-    public static ShapeQuery.Builder shape() {
-        return new ShapeQuery.Builder();
-    }
-
-    /**
-     * Creates a builder for the {@link XyShapeQuery shape} {@code Query} variant.
+     * Creates a builder for the {@link XyShapeQuery xy_shape} {@code Query} variant.
      */
     public static XyShapeQuery.Builder xyShape() {
         return new XyShapeQuery.Builder();