+++ java-client/src/main/java/org/opensearch/client/opensearch/ingest/ShapeType.java
@@ -39,8 +39,6 @@ import org.opensearch.client.json.JsonpDeserializable;
 public enum ShapeType implements JsonEnum {
     GeoShape("geo_shape"),
 
-    Shape("shape"),
-
     XyShape("xy_shape"),
 
     ;